requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-image-cropping==0.3.7
django-image-optimizer==1.0.0
django-widget-tweaks==1.4.5
easy-thumbnails==2.7
idna==2.8
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
tinify==1.5.1
urllib3==1.25.8
